requirements.txt
asgiref==3.8.1
cloudinary
django-cloudinary-storage
dj-database-url==3.0.1
Django==5.2.1
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
pillow==11.2.1
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.3
gunicorn==21.2.0
whitenoise==6.11.0
